requirements.txt
pandas
numpy
xgboost
scikit-learn
folium
streamlit
streamlit-folium
joblib
matplotlib
requests
python-dotenv